refactor: move into core
